鼠标做了基本测试，存在问题，鼠标移动y是反的，另外server可能会一次收到2条以上鼠标信息

+++ Client/Assets/Scripts/client/ui/Mouse/MousePanelMidArea.cs
@@ -22,8 +22,8 @@ public class MousePanelMidArea : MonoBehaviour,IDragHandler
     {
         Default.SB.Clear();
         Default.SB.Append("mouse|mid|");
-        int offy = deltaY * Default.MOUSE_MID_SPEED;
-        Default.SB.Append(offy);
+        float offy = deltaY * Default.MOUSE_MID_SPEED;
+        Default.SB.Append((int)offy);
         string key = Default.SB.ToString();
         SimpleEventMgr.Send(Default.EVENT_MOUSE, key);
     }
